fix: change passRate from float to double to avoid precision loss
float only has ~7 significant digits, causing 99.99965% to round up
to 100.0% and making the assert condition silently pass.

+++ src/main/java/custom/components/HybridSearchComp.java
@@ -397,7 +397,7 @@ public class HybridSearchComp {
         log.info("TP50:{}", MathUtil.calculateTP99(costTimeTotal, 0.50f));
 
         commonResult = CommonResult.builder().result(ResultEnum.SUCCESS.result).build();
-        float passRate = (float) (100.0 * successNum / requestNum);
+        double passRate = 100.0 * successNum / requestNum;
         // assertions
         List<String> assertMessages = new ArrayList<>();
         if (requestNum == 0) {
