+++ src/main/java/custom/entity/result/QueryResult.java
@@ -11,7 +11,7 @@ public class QueryResult {
     CommonResult commonResult;
     int concurrencyNum;
     long requestNum;
-    float passRate;
+    double passRate;
     float costTime;
     float rps;
     double avg;
